New L&F, universal icon, python deprecated and repo cleanup
The python version has been deprecated since it was harder and slower to work with, plus it was broken.
The L&F has been changed to that of the host system.
The icon has been fixed so now it holds on all screens.
Finally the repo has been cleaned up with the elimination of the python version and addition of a new .gitignore file